+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # DMP Tool Apollo Server Change Log
 
 ### Added
+- Added some logic to the `prepareObjectForLogs` function to redact sensitive info
 - Added `redact` to the pino logger to prevent sensitive information from being logged
 - Added resolver to handle resending of invite for project collaborators [#433]
 - Added bash to AWS Dockerfile for session manager